Major refactor after Saturday classes
