+++ app/entities/token.py
@@ -1,0 +1,12 @@
+from typing import Union
+
+from pydantic import BaseModel
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+
+class TokenData(BaseModel):
+    username: Union[str, None] = None
