+++ app/entities/user.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class User(BaseModel):
+    username: str
+
+
+class UserInDB(User):
+    hashed_password: str
