--- app/src/image_inverter/base_inverter.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-
-
-class BaseImageInverter(ABC):
-
-    def __init__(self):
-        pass
-
-    @abstractmethod
-    def invert(self, img_data: bytes):
-        pass
